[cleanup] Use a single entry point and rename some modules
- use only run.py entry point
- rename app / flask app
