+++ auth/cli.py
@@ -4,7 +4,7 @@ import json
 
 import click
 
-import auth
+import auth.flask_app
 
 
 HTTP_PORT = 5000
@@ -24,11 +24,11 @@ def run_auth(config, authdata):
     with open(config) as f:
         config_content = json.load(f)
 
-    auth.configure(config_content, authdata)
+    auth.flask_app.configure(config_content, authdata)
 
     click.launch(f"http://127.0.0.1:{HTTP_PORT}/")
 
-    auth.app.run(port=HTTP_PORT, debug=True)
+    auth.flask_app.app.run(port=HTTP_PORT, debug=True)
 
 
 if __name__ == '__main__':
