+++ run.py
@@ -6,7 +6,11 @@ import os
 
 import click
 
-import app
+import auth.flask_app
+import generator
+
+
+HTTP_PORT = 5000
 
 
 @click.command()
@@ -14,24 +18,41 @@ import app
               metavar="JSON_FILE", type=click.Path(exists=True))
 @click.option("-a", "--authdata", default="account.json",
               metavar="JSON_FILE", type=click.Path(exists=True))
+@click.option("-r", "--register", is_flag=True,
+              help="Register Strava account.")
+@click.option("-s", "--sync", is_flag=True,
+              help="Sync activities.")
+
 @click.option("-p", "--pois",
               metavar="JSON_FILE", type=click.Path())
 @click.option("-d", "--data", default="data.db",
               metavar="DATA_FILE", type=click.Path())
 @click.option("-o", "--output", default="web/activities.js",
               metavar="JS_FILE", type=click.Path())
-@click.option("-s", "--sync", is_flag=True,
-              help="Sync activities.")
 @click.option("-b", "--browser", is_flag=True,
               help="Open the generated website in a web browser.")
 @click.option("-f", "--force", is_flag=True,
               help="Force sync for older activities than the last synced.")
-def run(config, authdata, pois, data, output, sync, browser, force):
+def run(config, authdata, pois, data, output, sync, browser, force, register):
+
+    if register:
+        """
+        Run a simple web server to get authentication data to run the sync process
+
+        Read from config.json file and output to account.json
+        """
+        with open(config) as f:
+            config_content = json.load(f)
+
+        auth.flask_app.configure(config_content, authdata)
+        click.launch(f"http://127.0.0.1:{HTTP_PORT}/")
+        auth.flask_app.app.run(port=HTTP_PORT, debug=True)
+
     # Drop DB if sync and force mode enabled
     if sync and force:
         os.remove(data)
 
-    main = app.main.Main(data)
+    main = generator.main.Main(data)
 
     with open(config) as f:
         main.set_strava_app_config(json.load(f))
